Unit tests for Aggregates - New instance with updates - Aggregate rehydratated by an event stream

+++ EventSourcingStoreManager/ICommand.cs
@@ -9,7 +9,7 @@ namespace BaseDomainObjects
     public interface ICommand
     {
         Guid Id { get; }
-        ulong Version { get; set; }
+        int Version { get; set; }
 
     }
 }
